x/kyc/types/message_create_kyc_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"kyc/testutil/sample"
)

func TestMsgCreateKyc_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateKyc
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateKyc{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateKyc{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
